pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask>=3.0.3",
    "flask-sqlalchemy>=3.1.1",
    "psycopg2-binary>=2.9.10",
    "groq>=0.11.0",
    "google-generativeai>=0.8.3",
    "together>=1.3.2",
    "supabase>=2.9.0",
    "google-cloud-texttospeech>=2.18.0",
    "gtts>=2.5.3",
    "eventlet>=0.37.0",
    "flask-socketio>=5.4.1",
    "python-socketio>=5.11.4",
    "python-engineio>=4.10.1",
    "hume~=0.4.2",
    "sounddevice>=0.5.1",
    "websockets>=12.0",
    "pypdf2>=3.0.1",
    "ebooklib>=0.18",
    "beautifulsoup4>=4.12.3",
    "nltk>=3.9.1",
    "werkzeug>=3.0.4",
    "sqlalchemy>=2.0.36",
    "requests>=2.32.3",
    "python-dotenv>=1.0.1",
    "langchain>=0.3.7",
    "langchain-core>=0.3.19",
    "langchain-community>=0.3.4",
    "langchain-google-genai>=2.0.5",
    "pillow>=10.0.0",
]
